data_collection/requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.5
Brotli==1.0.9
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
fake-useragent==0.1.11
h11==0.13.0
h2==4.1.0
hpack==4.0.0
httmock==1.4.0
http-request-randomizer==1.3.2
hyperframe==6.0.1
idna==3.4
kaitaistruct==0.10
lxml==4.9.1
outcome==1.2.0
proxy-randomizer==1.3.3
psutil==5.9.2
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
selenium==4.4.3
selenium-wire==4.6.5
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.2.0
zstandard==0.18.0
